migrate formatKeyboardShortcut to utils
